+++ src/Console/ModelsCommand.php
@@ -37,6 +37,7 @@ use Illuminate\Database\Eloquent\Relations\MorphToMany;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Database\Schema\Builder;
 use Illuminate\Filesystem\Filesystem;
+use Illuminate\Support\Arr;
 use Illuminate\Support\Collection;
 use Illuminate\Support\Str;
 use phpDocumentor\Reflection\Types\ContextFactory;
@@ -528,6 +529,7 @@ class ModelsCommand extends Command
                     case 'varchar':
                     case 'bpchar':
                     case 'nvarchar':
+                    case 'binary':
                         $type = 'string';
                         break;
                     case 'integer':
@@ -600,10 +602,7 @@ class ModelsCommand extends Command
             // methods that resemble mutators but aren't.
             $reflections = array_filter($reflections, function (\ReflectionMethod $methodReflection) {
                 return !$methodReflection->isPrivate() && !(
-                    in_array(
-                        \Illuminate\Database\Eloquent\Concerns\HasAttributes::class,
-                        $methodReflection->getDeclaringClass()->getTraitNames()
-                    ) && (
+                    $methodReflection->getDeclaringClass()->getName() === Model::class && (
                         $methodReflection->getName() === 'setClassCastableAttribute' ||
                         $methodReflection->getName() === 'setEnumCastableAttribute'
                     )
@@ -629,18 +628,15 @@ class ModelsCommand extends Command
                         $this->setProperty($name, $type, true, null, $comment);
                     }
                 } elseif ($isAttribute) {
-                    $name = Str::snake($method);
-                    $types = $this->getAttributeReturnType($model, $reflection);
-                    $comment = $this->getCommentFromDocBlock($reflection);
-
-                    if ($types->has('get')) {
-                        $type = $this->getTypeInModel($model, $types['get']);
-                        $this->setProperty($name, $type, true, null, $comment);
-                    }
-
-                    if ($types->has('set')) {
-                        $this->setProperty($name, null, null, true, $comment);
-                    }
+                    $types = $this->getAttributeTypes($model, $reflection);
+                    $type = $this->getTypeInModel($model, $types->get('get') ?: $types->get('set')) ?: null;
+                    $this->setProperty(
+                        Str::snake($method),
+                        $type,
+                        $types->has('get'),
+                        $types->has('set'),
+                        $this->getCommentFromDocBlock($reflection)
+                    );
                 } elseif (
                     Str::startsWith($method, 'set') && Str::endsWith(
                         $method,
@@ -653,7 +649,7 @@ class ModelsCommand extends Command
                         $comment = $this->getCommentFromDocBlock($reflection);
                         $this->setProperty($name, null, null, true, $comment);
                     }
-                } elseif (Str::startsWith($method, 'scope') && $method !== 'scopeQuery') {
+                } elseif (Str::startsWith($method, 'scope') && $method !== 'scopeQuery' && $method !== 'scope' && $method !== 'scopes') {
                     //Magic set<name>Attribute
                     $name = Str::camel(substr($method, 5));
                     if (!empty($name)) {
@@ -824,12 +820,17 @@ class ModelsCommand extends Command
         $fkProp = $reflectionObj->getProperty('foreignKey');
         $fkProp->setAccessible(true);
 
-        if ($relation === 'belongsTo') {
-            return isset($this->nullableColumns[$fkProp->getValue($relationObj)]) ||
-                !in_array($fkProp->getValue($relationObj), $this->foreignKeyConstraintsColumns, true);
+        foreach (Arr::wrap($fkProp->getValue($relationObj)) as $foreignKey) {
+            if (isset($this->nullableColumns[$foreignKey])) {
+                return true;
+            }
+
+            if (!in_array($foreignKey, $this->foreignKeyConstraintsColumns, true)) {
+                return true;
+            }
         }
 
-        return isset($this->nullableColumns[$fkProp->getValue($relationObj)]);
+        return false;
     }
 
     /**
@@ -1167,7 +1168,10 @@ class ModelsCommand extends Command
         return $this->laravel['config']->get('ide-helper.model_camel_case_properties', false);
     }
 
-    protected function getAttributeReturnType(Model $model, \ReflectionMethod $reflectionMethod): Collection
+    /**
+     * @psalm-suppress NoValue
+     */
+    protected function getAttributeTypes(Model $model, \ReflectionMethod $reflectionMethod): Collection
     {
         // Private/protected ReflectionMethods require setAccessible prior to PHP 8.1
         $reflectionMethod->setAccessible(true);
@@ -1175,13 +1179,25 @@ class ModelsCommand extends Command
         /** @var Attribute $attribute */
         $attribute = $reflectionMethod->invoke($model);
 
-        return collect([
-            'get' => $attribute->get ? optional(new \ReflectionFunction($attribute->get))->getReturnType() : null,
-            'set' => $attribute->set ? optional(new \ReflectionFunction($attribute->set))->getReturnType() : null,
-        ])
-            ->filter()
+        $methods = new Collection();
+
+        if ($attribute->get) {
+            $methods['get'] = optional(new \ReflectionFunction($attribute->get))->getReturnType();
+        }
+        if ($attribute->set) {
+            $function = optional(new \ReflectionFunction($attribute->set));
+            if ($function->getNumberOfParameters() === 0) {
+                $methods['set'] = null;
+            } else {
+                $methods['set'] = $function->getParameters()[0]->getType();
+            }
+        }
+
+        return $methods
             ->map(function ($type) {
-                if ($type instanceof \ReflectionUnionType) {
+                if ($type === null) {
+                    $types = collect([]);
+                } elseif ($type instanceof \ReflectionUnionType) {
                     $types = collect($type->getTypes())
                         /** @var ReflectionType $reflectionType */
                         ->map(function ($reflectionType) {
@@ -1192,7 +1208,7 @@ class ModelsCommand extends Command
                     $types = collect($this->extractReflectionTypes($type));
                 }
 
-                if ($type->allowsNull()) {
+                if ($type && $type->allowsNull()) {
                     $types->push('null');
                 }
 
@@ -1442,8 +1458,7 @@ class ModelsCommand extends Command
     {
         $reflection = $model instanceof ReflectionClass
             ? $model
-            : new ReflectionObject($model)
-        ;
+            : new ReflectionObject($model);
 
         $className = trim($className, '\\');
         $writingToExternalFile = !$this->write || $this->write_mixin;